FIX: fix norm
