+++ includes/parse_map.h
@@ -6,7 +6,7 @@
 /*   By: hmouis <hmouis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/01 06:28:38 by hmouis            #+#    #+#             */
-/*   Updated: 2025/08/02 17:44:08 by hmouis           ###   ########.fr       */
+/*   Updated: 2025/09/09 12:02:18 by hmouis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 #include "./include.h"
 
 
-
+void        get_player_position(char **map, int *player_x, int *player_y);
 bool        file_name(char *str);
 void        fill_struct_info(t_map_info *info, t_info **final_info); 
 int         count_lines(int *fd, char *file_name);
